[OPP-1331] fjerne og rydde i all koden
Teamet ble enige om ktor for login-appen, og nginx for frontend-appen.
Rydder derfor bort mye av duplikat-koden, og fjerner unøvendig kompleksitet fra repoet (e.g common biten og sånt).
