+++ .github/workflows/login-app-workflow.yml
@@ -4,7 +4,7 @@ on:
   push:
     paths-ignore:
       - 'oidc-stub/**'
-      - 'frontend-app/**'
+      - 'frontend-image/**'
       - '**.md'
       - '.editorconfig'
       - '.gitignore'
