+++ docker-compose.yml
@@ -2,74 +2,36 @@ version: '3.4'
 services:
   oidc-stub:
     build:
-      context: .
-      dockerfile: oidc-stub/Dockerfile
-      network: host
+      context: oidc-stub
     ports:
       - 8080:8080
     environment:
       DOCKER_COMPOSE: "true"
   modialogin:
     build:
-      context: .
-      dockerfile: login-app/Dockerfile
-      network: host
-    ports:
-      - 8081:8080
-    depends_on:
-      - oidc-stub
-    environment:
-      APPNAME: "modialogin"
-      IDP_DISCOVERY_URL: "http://oidc-stub:8080/.well-known/openid-configuration"
-      IDP_CLIENT_ID: "modia-p"
-      IDP_CLIENT_SECRET: "modia-p"
-      X_FORWARDING_PORT: "8081"
-      DOCKER_COMPOSE: "true"
-  modiafrontend:
-    build:
-      context: .
-      dockerfile: frontend-app/Dockerfile
-      network: host
+      context: login-app
     ports:
       - 8082:8080
     depends_on:
       - oidc-stub
-      - modialogin
     environment:
-      APPNAME: "modiafrontend"
       IDP_DISCOVERY_URL: "http://oidc-stub:8080/.well-known/openid-configuration"
-      IDP_CLIENT_ID: "modia-p"
-      DELEGATED_LOGIN_URL: "http://localhost:8081/modialogin/api/start"
+      IDP_CLIENT_ID: "foo"
+      IDP_CLIENT_SECRET: "bar"
       X_FORWARDING_PORT: "8082"
       DOCKER_COMPOSE: "true"
-  nginxfrontend:
+  frontend:
     build:
-      context: nginx-frontend-app
-      network: host
+      context: frontend-image
     ports:
-      - 8085:8085
+      - 8083:8080
     depends_on:
       - oidc-stub
       - modialogin
     environment:
-      APP_NAME: "nginxfrontend"
+      APP_NAME: "frontend"
       APP_VERSION: "localhost"
       IDP_DISCOVERY_URL: "http://oidc-stub:8080/.well-known/openid-configuration"
-      IDP_CLIENT_ID: "modia-p"
-      DELEGATED_LOGIN_URL: "http://localhost:8086/nginxlogin/api/start"
+      IDP_CLIENT_ID: "foo"
+      DELEGATED_LOGIN_URL: "http://localhost:8082/modialogin/api/start"
       AUTH_TOKEN_RESOLVER: "cookie:modia_ID_token"
-  nginxlogin:
-    build:
-      context: nginx-login-app
-      network: host
-    ports:
-      - 8086:8086
-    depends_on:
-      - oidc-stub
-    environment:
-      APP_NAME: "nginxlogin"
-      APP_VERSION: "localhost"
-      IDP_DISCOVERY_URL: "http://oidc-stub:8080/.well-known/openid-configuration"
-      IDP_CLIENT_ID: "modia-p"
-      IDP_CLIENT_SECRET: "modia-p"
-
