--- frontend-app/Dockerfile
@@ -1,4 +1,0 @@
-FROM docker.pkg.github.com/navikt/pus-nais-java-app/pus-nais-java-app:java11
-
-COPY frontend-app/index.html /app
-COPY frontend-app/build/libs/app.jar app.jar
