--- frontend-app/build.gradle.kts
@@ -1,72 +1,0 @@
-import org.gradle.jvm.tasks.Jar
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-val ktorVersion = "1.5.2"
-val logbackVersion = "1.2.3"
-val logstashVersion = "5.1"
-
-plugins {
-    kotlin("jvm") version "1.4.30"
-    kotlin("plugin.serialization") version "1.4.30"
-}
-
-group = "no.nav"
-version = ""
-
-repositories {
-    mavenCentral()
-    jcenter()
-}
-
-dependencies {
-    implementation(kotlin("stdlib"))
-    implementation("io.ktor:ktor-server-netty:$ktorVersion")
-    implementation("io.ktor:ktor-auth:$ktorVersion")
-    implementation("io.ktor:ktor-auth-jwt:$ktorVersion")
-    implementation("io.ktor:ktor-serialization:$ktorVersion")
-    implementation("io.ktor:ktor-client:$ktorVersion")
-    implementation("io.ktor:ktor-client-cio:$ktorVersion")
-    implementation("io.ktor:ktor-client-auth:$ktorVersion")
-    implementation("io.ktor:ktor-client-serialization:$ktorVersion")
-    implementation("io.ktor:ktor-client-logging:$ktorVersion")
-    implementation("ch.qos.logback:logback-classic:$logbackVersion")
-    implementation("net.logstash.logback:logstash-logback-encoder:$logstashVersion")
-    implementation("dev.nohus:AutoKonfig:1.0.0")
-    implementation(project(":common"))
-
-    testImplementation("org.junit.jupiter:junit-jupiter:5.6.2")
-    testImplementation("com.nimbusds:nimbus-jose-jwt:9.7")
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-}
-tasks.withType<KotlinCompile> {
-    kotlinOptions.jvmTarget = "11"
-}
-
-tasks.test {
-    useJUnitPlatform()
-    testLogging {
-        events("passed", "skipped", "failed")
-    }
-}
-
-val fatJar = task("fatJar", type = Jar::class) {
-    archiveBaseName.set("app")
-    duplicatesStrategy = DuplicatesStrategy.INCLUDE
-    manifest {
-        attributes["Implementation-Title"] = "ModiaLogin"
-        attributes["Implementation-Version"] = archiveVersion
-        attributes["Main-Class"] = "no.nav.modialogin.MainKt"
-    }
-    from(configurations.runtimeClasspath.get().map({ if (it.isDirectory) it else zipTree(it) }))
-    with(tasks.jar.get() as CopySpec)
-}
-
-tasks {
-    "build" {
-        dependsOn(fatJar)
-    }
-}
