+++ frontend-image/Dockerfile
@@ -11,7 +11,7 @@ RUN luarocks install lua-resty-session \
     && luarocks install lua-resty-openidc
 
 COPY nginx.nginx          /etc/nginx/conf.d/nginx.conf.template
-COPY auth-flow.lua       /usr/local/openresty/nginx/
+COPY oidc_protected.lua   /usr/local/openresty/nginx/
 COPY start-nginx.sh       /usr/sbin/start-nginx
 RUN chmod u+x /usr/sbin/start-nginx
 RUN mkdir -p /nginx
@@ -19,7 +19,8 @@ RUN mkdir -p /app
 
 # Only used to test loading of secrets, should be commented out.
 COPY testsecrets.env /var/run/secrets/nais.io/vault/
+COPY default-index.html /app/index.html
 
-EXPOSE 8086
+EXPOSE 8080
 
 CMD ["start-nginx"]
