+++ frontend-image/nginx.nginx
@@ -13,13 +13,15 @@ resolver_timeout 3s;
 
 server {
     server_name _;
-    listen 8086;
+    listen 8080;
+    root "/app";
 
     set $app_name               "${APP_NAME}";
     set $app_version            "${APP_VERSION}";
     set $idp_discovery_url      "${IDP_DISCOVERY_URL}";
     set $idp_client_id          "${IDP_CLIENT_ID}";
-    set $idp_client_secret      "${IDP_CLIENT_SECRET}";
+    set $delegated_login_url    "${DELEGATED_LOGIN_URL}";
+    set $auth_token_location    "${AUTH_TOKEN_RESOLVER}";
 
     location = "/${APP_NAME}/internal/isAlive" {
         default_type text/plan;
@@ -36,16 +38,23 @@ server {
         return 200 "Application: ${APP_NAME}\nVersion: ${APP_VERSION}";
     }
 
-    location = "/${APP_NAME}/api/start" {
-        access_by_lua_file auth-flow.lua;
+
+    location = / {
+        return 301 "${APP_NAME}/";
     }
-    location = "/${APP_NAME}/api/login" {
-        access_by_lua_file auth-flow.lua;
+    location "/${APP_NAME}" {
+        return 301 "${APP_NAME}/";
     }
-    location = "/${APP_NAME}/api/refresh" {
-        access_by_lua_file auth-flow.lua;
+    location "/${APP_NAME}/" {
+        access_by_lua_file oidc_protected.lua;
+        try_files $uri @rewrites;
     }
-    location = "/${APP_NAME}/api/complete" {
-        access_by_lua_file auth-flow.lua;
+
+    # No matching file, send to default-index.html as we expect it to handle this uri.
+    location @rewrites {
+        index index.html;
+        try_files $uri $uri/ /index.html;
     }
+
+    include /nginx/*.nginx;
 }
