+++ frontend-image/oidc_protected.lua
