+++ frontend-image/testsecrets.env
@@ -1,0 +1,1 @@
+DUMMY_SECRET_FOR_TESTING_ONLY="This is simply here to verify that env-loading of secrets work. Please ignore."
