--- nginx-frontend-app/Dockerfile
@@ -1,26 +1,0 @@
-FROM openresty/openresty:alpine-fat
-
-# User env var is needed for luarocks to not complain.
-ENV APP_DIR="/app" \
-    USER="root"
-
-RUN apk add --no-cache --update bash
-RUN luarocks install lua-resty-session \
-    && luarocks install lua-resty-http \
-    && luarocks install lua-resty-jwt \
-    && luarocks install lua-resty-openidc
-
-COPY nginx.nginx          /etc/nginx/conf.d/nginx.conf.template
-COPY oidc_protected.lua   /usr/local/openresty/nginx/
-COPY start-nginx.sh       /usr/sbin/start-nginx
-RUN chmod u+x /usr/sbin/start-nginx
-RUN mkdir -p /nginx
-RUN mkdir -p /app
-
-# Only used to test loading of secrets, should be commented out.
-COPY testsecrets.env /var/run/secrets/nais.io/vault/
-COPY default-index.html /app/index.html
-
-EXPOSE 8085
-
-CMD ["start-nginx"]
