--- nginx-frontend-app/default-index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Modialogin dummy side</title>
-</head>
-<body>
-<h1>Modialogin dummy side</h1>
-</body>
-</html>
