--- nginx-frontend-app/nginx.nginx
@@ -1,60 +1,0 @@
-charset utf-8;
-error_log    /dev/stdout  info;
-
-lua_package_path '~/lua/?.lua;;';
-lua_shared_dict discovery 1m;
-lua_shared_dict introspection 1m;
-lua_shared_dict jwks 1m;
-lua_code_cache on; # Need to be set for session storage to work.
-lua_use_default_type off;
-
-resolver "${RESOLVER}" ipv6=off;
-resolver_timeout 3s;
-
-server {
-    server_name _;
-    listen 8085;
-    root "/app";
-
-    set $app_name               "${APP_NAME}";
-    set $app_version            "${APP_VERSION}";
-    set $idp_discovery_url      "${IDP_DISCOVERY_URL}";
-    set $idp_client_id          "${IDP_CLIENT_ID}";
-    set $delegated_login_url    "${DELEGATED_LOGIN_URL}";
-    set $auth_token_location    "${AUTH_TOKEN_RESOLVER}";
-
-    location = "/${APP_NAME}/internal/isAlive" {
-        default_type text/plan;
-        return 200 "Alive";
-    }
-
-    location = "/${APP_NAME}/internal/isReady" {
-        default_type text/plan;
-        return 200 "Ready";
-    }
-
-    location = "/${APP_NAME}/internal/selftest" {
-        default_type text/plan;
-        return 200 "Application: ${APP_NAME}\nVersion: ${APP_VERSION}";
-    }
-
-
-    location = / {
-        return 301 "${APP_NAME}/";
-    }
-    location "/${APP_NAME}" {
-        return 301 "${APP_NAME}/";
-    }
-    location "/${APP_NAME}/" {
-        access_by_lua_file oidc_protected.lua;
-        try_files $uri @rewrites;
-    }
-
-    # No matching file, send to default-index.html as we expect it to handle this uri.
-    location @rewrites {
-        index index.html;
-        try_files $uri $uri/ /index.html;
-    }
-
-    include /nginx/*.nginx;
-}
