--- nginx-login-app/testsecrets.env
@@ -1,1 +1,0 @@
-TEST_SECRET=abligoyer
