+++ oidc-stub/Dockerfile
@@ -1,2 +1,2 @@
 FROM docker.pkg.github.com/navikt/pus-nais-java-app/pus-nais-java-app:java11
-COPY oidc-stub/build/libs/app.jar app.jar
+COPY build/libs/app.jar app.jar
