+++ settings.gradle.kts
@@ -1,5 +1,3 @@
 rootProject.name = "modialogin"
-include("common")
 include("login-app")
-include("frontend-app")
 include("oidc-stub")
